mybatis

+++ mybatis/src/main/java/com/yangxvhao/demo/model/Person.java
@@ -9,6 +9,6 @@ import lombok.Data;
 @Data
 public class Person {
     private String name;
-    private int age;
+    private Long age;
     private Address address;
 }
